+++ apps/web/src/app/shared/components/vertical-stepper/vertical-step-content.component.ts
