+++ apps/web/src/app/shared/components/vertical-stepper/vertical-step.component.html
