+++ apps/web/src/app/shared/components/vertical-stepper/vertical-stepper.component.ts
